+++ example/lib/Views/mini_view.dart
@@ -88,3 +88,48 @@ class MiniView extends StatelessWidget {
     print('this listener called when widget of mcNum rebuild');
   }
 }
+
+// class McMiniViewExample extends StatelessWidget {
+//   // use mini for convert value to McValue
+//   final McValue<String> myStringValue = "My Value".mini;
+//   final McValue<int> myIntValue = 2021.mini;
+//   @override
+//   Widget build(BuildContext context) {
+//     return Scaffold(
+//       body: Container(
+//         //use your value in McMV and if value changed will rebuild widget for show your new value
+//         child: Column(
+//           mainAxisAlignment: MainAxisAlignment.center,
+//           children: [
+//             // use McValue for every widget
+//             McMV(myStringValue, () => Text(myStringValue.v)),
+//             McMV(myStringValue, () => Text(myIntValue.v.toString())),
+//             const SizedBox(
+//               height: 25.0,
+//             ),
+//             // merge multi MCValue in one widget
+//             McMV(McValue.merge([myStringValue, myIntValue]), () {
+//               return Row(
+//                 mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+//                 children: [
+//                   Text(myStringValue.v),
+//                   Text(myIntValue.v.toString())
+//                 ],
+//               );
+//             })
+//           ],
+//         ),
+//       ),
+//       floatingActionButton: FloatingActionButton(
+//         backgroundColor: Theme.of(context).primaryColor,
+//         onPressed: () {
+//           // change value
+//           myStringValue.v = "Value Changed";
+//           myIntValue.v = 2022;
+//         },
+//         tooltip: 'change Value',
+//         child: Icon(Icons.change_circle),
+//       ),
+//     );
+//   }
+// }