+++ example/lib/Views/post_View.dart
@@ -7,66 +7,97 @@ class PostExample extends StatelessWidget {
   // readOnly means if you close and open this screen you will use same data without update it from Api
   // [mc] is instance of Mccontroller injected in Object by extension for use it easily anywhere
   final Post post = McController().add<Post>('posts', Post(), readOnly: true);
-  final McRequest rq = McController().get<McRequest>("rq");
+  // get request by key
+  final McRequest request = McController().get<McRequest>("rq");
   PostExample({this.title});
   final String title;
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        title: Text(
-          "Refresh Posts with swip to down or from here =>",
-          style: TextStyle(fontSize: 11.0),
+        appBar: AppBar(
+          title: Text(
+            "Refresh Posts with swip to down or from here =>",
+            style: TextStyle(fontSize: 11.0),
+          ),
+          actions: [
+            IconButton(
+                icon: Icon(Icons.data_usage),
+                // Refresh Data from Api
+                onPressed: () => refresh())
+          ],
         ),
-        actions: [
-          IconButton(
-              icon: Icon(Icons.data_usage),
-              // Refresh Data from Api
-              onPressed: () => refresh())
-        ],
-      ),
-      body: Container(
+        body: Container(
           height: MediaQuery.of(context).size.height,
           width: MediaQuery.of(context).size.width,
-          child: McView(
-            call: () => rq.getObjData("posqts", post, multi: true),
-            model: post,
-            onError: (McException exception) {
-              return Column(
-                children: [Text(exception.exception), Text(exception.response)],
-              );
-            },
-            // call api if model is empty
-            callType: CallType.callIfModelEmpty,
-            builder: (context) {
-              return RefreshIndicator(
-                onRefresh: () {
-                  return refresh();
+          child: RefreshIndicator(
+              onRefresh: () {
+                return refresh();
+              },
+              child: McView(
+                // call api method
+                call: () => request.getObjData("posts", post, multi: true),
+                // your model generated
+                model: post,
+                // handle errors
+                onError: (McException exception,Function() reload) {
+                  return Center(
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: [
+                        Text(exception.exception),
+                        Text(exception.response),
+                        TextButton(onPressed: reload, child: Text("retry"))
+                      ],
+                    ),
+                  );
                 },
-                child: Container(
-                  height: MediaQuery.of(context).size.height * 0.852,
-                  child: ListView.builder(
-                    itemCount: post.multi.length,
-                    itemBuilder: (BuildContext context, int index) {
-                      return ListTile(
-                        leading: Text(post.multi[index].id.toString()),
-                        title: Text(post.multi[index].title),
-                        onTap: () => Navigator.of(context).push(
-                            MaterialPageRoute(builder: (BuildContext context) {
-                          return Details(index);
-                        })),
-                      );
-                    },
-                  ),
-                ),
-              );
-            },
-          )),
-    );
+                // call api if model is empty & you can choose another ways like default way asFuture(call once) & asStream (call every //[secondsOfStream] seconds)
+                callType: CallType.callIfModelEmpty,
+                // or
+                // callType: CallType.callAsStream,
+                // secondsOfStream: 1,
+                // customized your loading (default widget is CircularProgressIndicator)
+                // loader:CustomLoading(),
+                builder: (context) {
+                  return Container(
+                    height: MediaQuery.of(context).size.height * 0.852,
+                    child: ListView.builder(
+                      itemCount: post.multi.length,
+                      itemBuilder: (BuildContext context, int index) {
+                        // your data saved in multi list as Post model
+                        Post currentPost = post.multi[index];
+                        return ListTile(
+                            leading: Text(currentPost.id.toString()),
+                            title: Text(currentPost.title),
+                            onTap: () => Navigator.of(context).push(
+                                  MaterialPageRoute(
+                                      builder: (BuildContext context) {
+                                    return Details(index);
+                                  }),
+                                ));
+                      },
+                    ),
+                  );
+                },
+              )),
+        ));
   }
 
   Future<dynamic> refresh() {
-    return rq.getObjData("posts", post, multi: true);
+    // use hrrp method you want (get,post,put) + ObjData if you used model in McView and you can use JsonData for get data directly from api
+    return request.getObjData(
+      // endpoint
+      "posts",
+      // your model
+      post,
+      // if you received data as List multi will be true & if data as map you not should to define multi its false as default
+      multi: true,
+      // parameters for send it with request
+      // params:{"key":"value"},
+      // inspect method for determine exact json use for generate your model in first step
+      // if your api send data directly without any supplement values you not should define it
+      // inspect:(data)=>data["response"]
+    );
   }
 }
 
@@ -89,124 +120,3 @@ class Details extends StatelessWidget {
     );
   }
 }
-
-///////## multi[true] (FutureBuilder)
-///
-///
-// class MyHomePage extends StatelessWidget {
-//   MyHomePage({this.title});
-//   final String title;
-//   final Post post = Post();
-//  @override
-// Widget build(BuildContext context) {
-//   return Scaffold(
-//     appBar: AppBar(
-//       title: Text(title),
-//     ),
-//     body: Container(
-//         height: MediaQuery.of(context).size.height,
-//         width: MediaQuery.of(context).size.width,
-//         child: FutureBuilder(
-//           future: request.getObjData("posts", post, multi: true),
-//           builder: () {
-//             return post.loading
-//                 ? Center(child: CircularProgressIndicator())
-//                 : ListView.builder(
-//                     itemCount: snp.data.length,
-//                     itemBuilder: (BuildContext context, int index) {
-//                       return ExpansionTile(
-//                           title: Text(snp.data[index].title),
-//                           children: [
-//                             SizedBox(height: 5.0),
-//                             Text(snp.data[index].body)
-//                           ]);
-//                     },
-//                   );
-//           },
-//         )),
-//   );
-// }
-// }
-
-///// with [McView]
-///// ## multi[true]
-///
-///
-///
-// class MyHomePage extends StatelessWidget {
-//   MyHomePage({this.title});
-//   final String title;
-//   final Post post = Post();
-//   @override
-//   Widget build(BuildContext context) {
-//     return Scaffold(
-//       floatingActionButton: FloatingActionButton(
-//         child:Icon(Icons.get_app),
-//         onPressed:()=>request.getObjData("posts", post,multi:true)
-//       ),
-//       appBar: AppBar(
-//         title: Text(title),
-//       ),
-//       body: Container(
-//           height: MediaQuery.of(context).size.height,
-//           width: MediaQuery.of(context).size.width,
-//           child: McView(
-//             model: post,
-//             builder: () {
-//               return ListView.builder(
-//                 itemCount: post.multi.length,
-//                 itemBuilder: (BuildContext context, int index) {
-//                   return ExpansionTile(
-//                     title:Text(post.multi[index].title),
-//                     children:[
-//                       SizedBox(height:5.0),
-//                       Text(post.multi[index].body)
-//                     ]
-//                   );
-//                 },
-//               );
-//             },
-//           )),
-//     );
-//   }
-// }
-
-// //// ## multi [false]
-
-// class MyHomePage extends StatelessWidget {
-//   final String title;
-//   MyHomePage({this.title});
-//   final String title;
-//   final Post post = Post();
-//   @override
-//   Widget build(BuildContext context) {
-//     return Scaffold(
-//       floatingActionButton: FloatingActionButton(
-//           child: Icon(Icons.get_app),
-//           onPressed: () => request.getObjData("posts/1", post)),
-//       appBar: AppBar(
-//         title: Text(title),
-//       ),
-//       body: Container(
-//           height: MediaQuery.of(context).size.height,
-//           width: MediaQuery.of(context).size.width,
-//           child: McView(
-//             model: post,
-//             builder: () {
-//               return ExpansionTile(
-//                   title: Text(post.title),
-//                   children: [SizedBox(height: 5.0), Text(post.body)]);
-//             },
-//           )),
-//     );
-//   }
-// }
-
-// as Request Json || Model
-// ## multi[false]
-// Json: request.getJsonData("posts/1").then((value) => print(value));
-// Model: request.getObjData("posts/1",yourModelHere).then((value) => print(value));
-// ## multi[true]
-// Json: request.getJsonData("posts",multi: true).then((value) => print(value));
-// Model: request.getObjData("posts",yourModelHere,multi: true).then((value) => print(value));
-